+++ components/net/sal_socket/src/sal_socket.c
@@ -570,7 +570,7 @@ int sal_accept(int socket, struct sockaddr *addr, socklen_t *addrlen)
     /* check the network interface socket operations */
     SAL_NETDEV_SOCKETOPS_VALID(sock->netdev, pf, accept);
 
-    new_socket = pf->skt_ops->accept((int) sock->user_data, addr, addrlen);
+    new_socket = pf->skt_ops->accept((int)(size_t)sock->user_data, addr, addrlen);
     if (new_socket != -1)
     {
         int retval;
@@ -598,7 +598,7 @@ int sal_accept(int socket, struct sockaddr *addr, socklen_t *addrlen)
         }
 
         /* socket structure user_data used to store the acquired new socket */
-        new_sock->user_data = (void *) new_socket;
+        new_sock->user_data = (void *)(size_t)new_socket;
 
         return new_sal_socket;
     }
@@ -664,13 +664,13 @@ int sal_bind(int socket, const struct sockaddr *name, socklen_t namelen)
                 return -1;
             }
             sock->netdev = new_netdev;
-            sock->user_data = (void *) new_socket;
+            sock->user_data = (void *)(size_t)new_socket;
         }
     }
 
     /* check and get protocol families by the network interface device */
     SAL_NETDEV_SOCKETOPS_VALID(sock->netdev, pf, bind);
-    return pf->skt_ops->bind((int) sock->user_data, name, namelen);
+    return pf->skt_ops->bind((int)(size_t)sock->user_data, name, namelen);
 }
 
 int sal_shutdown(int socket, int how)
@@ -686,7 +686,7 @@ int sal_shutdown(int socket, int how)
     /* check the network interface socket opreation */
     SAL_NETDEV_SOCKETOPS_VALID(sock->netdev, pf, shutdown);
 
-    if (pf->skt_ops->shutdown((int) sock->user_data, how) == 0)
+    if (pf->skt_ops->shutdown((int)(size_t)sock->user_data, how) == 0)
     {
 #ifdef SAL_USING_TLS
         if (SAL_SOCKOPS_PROTO_TLS_VALID(sock, closesocket))
@@ -719,7 +719,7 @@ int sal_getpeername(int socket, struct sockaddr *name, socklen_t *namelen)
     /* check the network interface socket opreation */
     SAL_NETDEV_SOCKETOPS_VALID(sock->netdev, pf, getpeername);
 
-    return pf->skt_ops->getpeername((int) sock->user_data, name, namelen);
+    return pf->skt_ops->getpeername((int)(size_t)sock->user_data, name, namelen);
 }
 
 int sal_getsockname(int socket, struct sockaddr *name, socklen_t *namelen)
@@ -733,7 +733,7 @@ int sal_getsockname(int socket, struct sockaddr *name, socklen_t *namelen)
     /* check the network interface socket opreation */
     SAL_NETDEV_SOCKETOPS_VALID(sock->netdev, pf, getsockname);
 
-    return pf->skt_ops->getsockname((int) sock->user_data, name, namelen);
+    return pf->skt_ops->getsockname((int)(size_t)sock->user_data, name, namelen);
 }
 
 int sal_getsockopt(int socket, int level, int optname, void *optval, socklen_t *optlen)
@@ -747,7 +747,7 @@ int sal_getsockopt(int socket, int level, int optname, void *optval, socklen_t *
     /* check the network interface socket opreation */
     SAL_NETDEV_SOCKETOPS_VALID(sock->netdev, pf, getsockopt);
 
-    return pf->skt_ops->getsockopt((int) sock->user_data, level, optname, optval, optlen);
+    return pf->skt_ops->getsockopt((int)(size_t)sock->user_data, level, optname, optval, optlen);
 }
 
 int sal_setsockopt(int socket, int level, int optname, const void *optval, socklen_t optlen)
@@ -793,7 +793,7 @@ int sal_setsockopt(int socket, int level, int optname, const void *optval, sockl
         return pf->skt_ops->setsockopt((int) sock->user_data, level, optname, optval, optlen);
     }
 #else
-    return pf->skt_ops->setsockopt((int) sock->user_data, level, optname, optval, optlen);
+    return pf->skt_ops->setsockopt((int)(size_t)sock->user_data, level, optname, optval, optlen);
 #endif /* SAL_USING_TLS */
 }
 
@@ -811,7 +811,7 @@ int sal_connect(int socket, const struct sockaddr *name, socklen_t namelen)
     /* check the network interface socket opreation */
     SAL_NETDEV_SOCKETOPS_VALID(sock->netdev, pf, connect);
 
-    ret = pf->skt_ops->connect((int) sock->user_data, name, namelen);
+    ret = pf->skt_ops->connect((int)(size_t)sock->user_data, name, namelen);
 #ifdef SAL_USING_TLS
     if (ret >= 0 && SAL_SOCKOPS_PROTO_TLS_VALID(sock, connect))
     {
@@ -838,7 +838,7 @@ int sal_listen(int socket, int backlog)
     /* check the network interface socket opreation */
     SAL_NETDEV_SOCKETOPS_VALID(sock->netdev, pf, listen);
 
-    return pf->skt_ops->listen((int) sock->user_data, backlog);
+    return pf->skt_ops->listen((int)(size_t)sock->user_data, backlog);
 }
 
 int sal_recvfrom(int socket, void *mem, size_t len, int flags,
@@ -868,10 +868,10 @@ int sal_recvfrom(int socket, void *mem, size_t len, int flags,
     }
     else
     {
-        return pf->skt_ops->recvfrom((int) sock->user_data, mem, len, flags, from, fromlen);
+        return pf->skt_ops->recvfrom((int)(size_t)sock->user_data, mem, len, flags, from, fromlen);
     }
 #else
-    return pf->skt_ops->recvfrom((int) sock->user_data, mem, len, flags, from, fromlen);
+    return pf->skt_ops->recvfrom((int)(size_t)sock->user_data, mem, len, flags, from, fromlen);
 #endif
 }
 
@@ -905,7 +905,7 @@ int sal_sendto(int socket, const void *dataptr, size_t size, int flags,
         return pf->skt_ops->sendto((int) sock->user_data, dataptr, size, flags, to, tolen);
     }
 #else
-    return pf->skt_ops->sendto((int) sock->user_data, dataptr, size, flags, to, tolen);
+    return pf->skt_ops->sendto((int)(size_t)sock->user_data, dataptr, size, flags, to, tolen);
 #endif
 }
 
@@ -957,7 +957,7 @@ int sal_socket(int domain, int type, int protocol)
             }
         }
 #endif
-        sock->user_data = (void *) proto_socket;
+        sock->user_data = (void *)(size_t)proto_socket;
         return sock->socket;
     }
     socket_delete(socket);
@@ -977,7 +977,7 @@ int sal_closesocket(int socket)
     /* valid the network interface socket opreation */
     SAL_NETDEV_SOCKETOPS_VALID(sock->netdev, pf, socket);
 
-    if (pf->skt_ops->closesocket((int) sock->user_data) == 0)
+    if (pf->skt_ops->closesocket((int)(size_t)sock->user_data) == 0)
     {
 #ifdef SAL_USING_TLS
         if (SAL_SOCKOPS_PROTO_TLS_VALID(sock, closesocket))
@@ -1012,7 +1012,7 @@ int sal_ioctlsocket(int socket, long cmd, void *arg)
     /* check the network interface socket opreation */
     SAL_NETDEV_SOCKETOPS_VALID(sock->netdev, pf, ioctlsocket);
 
-    return pf->skt_ops->ioctlsocket((int) sock->user_data, cmd, arg);
+    return pf->skt_ops->ioctlsocket((int)(size_t)sock->user_data, cmd, arg);
 }
 
 #ifdef SAL_USING_POSIX
@@ -1020,7 +1020,7 @@ int sal_poll(struct dfs_fd *file, struct rt_pollreq *req)
 {
     struct sal_socket *sock;
     struct sal_proto_family *pf;
-    int socket = (int) file->fnode->data;
+    int socket = (int)(size_t)file->fnode->data;
 
     /* get the socket object by socket descriptor */
     SAL_SOCKET_OBJ_GET(sock, socket);
